Minify site.css + ads-interim.js to fix SEMrush "unminified CSS/JS" warning
SEMrush flagged /assets/site.css (and /assets/ads-interim.js) as unminified
across all pages that load them. Both are now minified via esbuild:
  - site.css        6848 → 2776 bytes (-59%)
  - ads-interim.js  2748 →  687 bytes (-75%)

Kept editable: the readable sources now live in assets/src/, and
scripts/minify-assets.js (npm run minify) rebuilds the served assets/<file>
from them — so the HTML needs no changes (same referenced paths) and the files
stay maintainable. Verified: all 42 CSS properties preserved (esbuild only
merges/optimises), every media query intact, JS still parses with its
disabled-by-default behaviour, and the page renders correctly.

esbuild added as a devDependency; package-lock.json deliberately not committed
(the Scope Guard workflow forbids it and no CI uses `npm ci`). node_modules is
gitignored.
